Menu layout
* Add Aleph menu layout
* Fix commentary issue in dark mode
* Fix highlighting issue for Chinese Bibles
