+++ .gitignore
@@ -7,3 +7,4 @@
 config.py
 **.log
 htmlResources/main.html
+htmlResources/study.html
