+++ Languages.py
@@ -137,8 +137,12 @@ class Languages:
         "menu1_wikiPages": "Wiki Pages",
         "menu1_update": "Update",
         "menu1_selectTheme": "Select Theme",
-        "menu1_default_theme": "Default theme",
-        "menu1_dark_theme": "Dark theme",
+        "menu1_default_theme": "Default Theme",
+        "menu1_dark_theme": "Dark Theme",
+        "menu1_selectMenuLayout": "Select Menu Layout",
+        "menu1_menuLayout": "Menu Layout",
+        "menu1_classic_menu_layout": "Classic Menu Layout",
+        "menu1_aleph_menu_layout": "Aleph Menu Layout",
         "menu1_selectDefaultLexicon": "Select Default Lexicon",
         "menu1_StrongsGreek": "Strong's Greek",
         "menu1_StrongsHebrew": "Strong's Hebrew",
@@ -146,6 +150,7 @@ class Languages:
         "menu1_setDefaultStrongsHebrewLexicon": "Set Default Strong's Hebrew Lexicon",
         "menu1_exit": "Exit",
         "menu2_view": "Layout",
+        "menu2_view": "Navigation",
         "menu2_all": "All Toolbars [Show / Hide]",
         "menu2_topOnly": "All Toolbars / Only Top Toolbar",
         "menu2_top": "Top Toolbar [Show / Hide]",
@@ -238,6 +243,62 @@ class Languages:
         "menu10_displayContent": "Display Content: Study / New Window",
         "menu10_bookOnNew": "Further book content will be opened on a new window.",
         "menu10_bookOnStudy": "Further book content will be opened on Study Window.",
+
+        "menu_abbreviations": "Abbreviations",
+        "menu_about": "About",
+        "menu_annotate": "Annotate",
+        "menu_apps": "Apps",
+        "menu_bible": "Bible",
+        "menu_bible_format": "Bible Format",
+        "menu_bible_chapter_notes": "Bible Chapter Notes",
+        "menu_bible_notes": "Bible Notes",
+        "menu_bible_verse_notes": "Bible Verse Notes",
+        "menu_chapter": "Chapter",
+        "menu_chineseFont": "Chinese Font",
+        "menu_command": "Command",
+        "menu_config_flags": "Set Config Flags",
+        "menu_data": "Data",
+        "menu_defaults": "Defaults",
+        "menu_display": "Display",
+        "menu_donate": "Donate",
+        "menu_edit_note": "Edit Note",
+        "menu_external_notes": "External Notes",
+        "menu_favouriteBible": "Favourite Bible",
+        "menu_font": "Font and Size",
+        "menu_font": "Font",
+        "menu_highlight": "Highlight",
+        "menu_history": "History",
+        "menu_language": "Language",
+        "menu_lexicon": "Lexicon",
+        "menu_library": "Library",
+        "menu_light_theme": "Light Theme",
+        "menu_main_page_down": "Main Page Scroll Down",
+        "menu_main_page_up": "Main Page Page Up",
+        "menu_main_scroll_to_top": "Main Scroll To Top",
+        "menu_navigation": "Navigation",
+        "menu_new_note": "New Note",
+        "menu_next_book": "Next Book",
+        "menu_notes": "Notes",
+        "menu_open_note": "Open Note",
+        "menu_previous_book": "Previous Book",
+        "menu_read_note": "Read Note",
+        "menu_reload": "Reload",
+        "menu_scroll": "Scroll",
+        "menu_search": "Search",
+        "menu_simple_formatted": "Simple/Formatted",
+        "menu_select_default_font": "Set Default Font and Size",
+        "menu_study_page_down": "Study Page Scroll Down",
+        "menu_study_page_up": "Study Page Page Up",
+        "menu_study_scroll_to_top": "Study Scroll To Top",
+        "menu_subheadings": "Subheadings",
+        "menu_tabs": "Tabs",
+        "menu_theme": "Theme",
+        "menu_toolbar": "Toolbar",
+        "menu_verse": "Verse",
+        "menu_verse_all_versions": "Verse All Versions",
+        "menu_wiki": "Wiki",
+        "menu_window": "Window",
+
         "context1_copy": "Copy",
         "context1_copy_html": "Copy HTML",
         "context1_copyReferences": "Copy All Bible References",
