+++ UniqueBibleAppVersion.txt
@@ -1,1 +1,1 @@
-19.4
+19.5
