+++ gui/ClassicMainWindow.py
@@ -1,0 +1,1058 @@
+import os, config, myTranslation
+from PySide2.QtGui import QIcon, Qt
+from PySide2.QtWidgets import (QAction, QToolBar, QPushButton, QLineEdit)
+from gui.MainWindow import MainWindow
+
+class ClassicMainWindow(MainWindow):
+
+    def __init__(self):
+        super().__init__()
+
+    def create_menu(self):
+        menu1 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu1_app"]))
+        screenSizeMenu = menu1.addMenu(config.thisTranslation["menu1_screenSize"])
+        screenSizeMenu.addAction(QAction(config.thisTranslation["menu1_fullScreen"], self, shortcut="Ctrl+S,F", triggered=self.fullsizeWindow))
+        screenSizeMenu.addAction(QAction(config.thisTranslation["menu1_smallSize"], self, shortcut="Ctrl+S,S", triggered=self.twoThirdWindow))
+        screenSizeMenu.addAction(QAction(config.thisTranslation["menu1_topHalf"], self, shortcut="Ctrl+S,1", triggered=self.topHalfScreenHeight))
+        screenSizeMenu.addAction(QAction(config.thisTranslation["menu1_bottomHalf"], self, shortcut="Ctrl+S,2", triggered=self.bottomHalfScreenHeight))
+        screenSizeMenu.addAction(QAction(config.thisTranslation["menu1_leftHalf"], self, shortcut="Ctrl+S,3", triggered=self.leftHalfScreenWidth))
+        screenSizeMenu.addAction(QAction(config.thisTranslation["menu1_rightHalf"], self, shortcut="Ctrl+S,4", triggered=self.rightHalfScreenWidth))
+        menu1.addSeparator()
+        menu1.addAction(QAction(config.thisTranslation["menu1_setDefaultFont"], self, triggered=self.setDefaultFont))
+        menu1.addAction(QAction(config.thisTranslation["menu1_setChineseFont"], self, triggered=self.setChineseFont))
+        menu1.addAction(QAction(config.thisTranslation["menu1_setAbbreviations"], self, triggered=self.setBibleAbbreviations))
+        menu1.addAction(QAction(config.thisTranslation["menu1_tabNo"], self, triggered=self.setTabNumberDialog))
+        menu1.addAction(QAction(config.thisTranslation["menu1_setMyFavouriteBible"], self, triggered=self.openFavouriteBibleDialog))
+        menu1.addAction(QAction(config.thisTranslation["menu1_setMyLanguage"], self, triggered=self.openMyLanguageDialog))
+        lexiconMenu = menu1.addMenu(config.thisTranslation["menu1_selectDefaultLexicon"])
+        lexiconMenu.addAction(QAction(config.thisTranslation["menu1_StrongsHebrew"], self, triggered=self.openSelectDefaultStrongsHebrewLexiconDialog))
+        lexiconMenu.addAction(QAction(config.thisTranslation["menu1_StrongsGreek"], self, triggered=self.openSelectDefaultStrongsGreekLexiconDialog))
+        themeMenu = menu1.addMenu(config.thisTranslation["menu1_selectTheme"])
+        themeMenu.addAction(QAction(config.thisTranslation["menu_light_theme"], self, triggered=self.setDefaultTheme))
+        themeMenu.addAction(QAction(config.thisTranslation["menu1_dark_theme"], self, triggered=self.setDarkTheme))
+        layoutMenu = menu1.addMenu(config.thisTranslation["menu1_selectMenuLayout"])
+        layoutMenu.addAction(
+            QAction(config.thisTranslation["menu1_classic_menu_layout"], self, triggered=self.setDefaultMenuLayout))
+        layoutMenu.addAction(QAction(config.thisTranslation["menu1_aleph_menu_layout"], self, triggered=self.setAlephMenuLayout))
+
+        menu1.addAction(QAction(config.thisTranslation["menu1_moreConfig"], self, triggered=self.moreConfigOptionsDialog))
+        menu1.addSeparator()
+        if (not self.isMyTranslationAvailable() and not self.isOfficialTranslationAvailable()) or (self.isMyTranslationAvailable() and not myTranslation.translationLanguage == config.userLanguage) or (self.isOfficialTranslationAvailable() and not config.translationLanguage == config.userLanguage):
+            menu1.addAction(QAction(config.thisTranslation["menu1_translateInterface"], self, triggered=self.translateInterface))
+        menu1.addAction(QAction(config.thisTranslation["menu1_toogleInterface"], self, triggered=self.toogleInterfaceTranslation))
+        menu1.addSeparator()
+        menu1.addAction(QAction(config.thisTranslation["menu1_update"], self, triggered=self.updateUniqueBibleApp))
+        menu1.addSeparator()
+        menu1.addAction(QAction(config.thisTranslation["menu1_remoteControl"], self, shortcut="Ctrl+O", triggered=self.manageRemoteControl))
+        menu1.addSeparator()
+        appIcon = QIcon(os.path.join("htmlResources", "theText.png"))
+        quit_action = QAction(appIcon, config.thisTranslation["menu1_exit"], self, shortcut="Ctrl+Q", triggered=qApp.quit)
+        menu1.addAction(quit_action)
+
+        menu2 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu2_view"]))
+        menu2.addAction(QAction(config.thisTranslation["menu2_all"], self, shortcut="Ctrl+J", triggered=self.setNoToolBar))
+        menu2.addAction(QAction(config.thisTranslation["menu2_topOnly"], self, shortcut="Ctrl+G", triggered=self.hideShowAdditionalToolBar))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_top"], self, triggered=self.hideShowMainToolBar))
+        menu2.addAction(QAction(config.thisTranslation["menu2_second"], self, triggered=self.hideShowSecondaryToolBar))
+        menu2.addAction(QAction(config.thisTranslation["menu2_left"], self, triggered=self.hideShowLeftToolBar))
+        menu2.addAction(QAction(config.thisTranslation["menu2_right"], self, triggered=self.hideShowRightToolBar))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_icons"], self, triggered=self.switchIconSize))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_landscape"], self, shortcut="Ctrl+L", triggered=self.switchLandscapeMode))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_study"], self, shortcut="Ctrl+W", triggered=self.parallel))
+        menu2.addAction(QAction(config.thisTranslation["menu2_bottom"], self, shortcut="Ctrl+T", triggered=self.cycleInstant))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_hover"], self, shortcut="Ctrl+=", triggered=self.enableInstantButtonClicked))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_larger"], self, shortcut="Ctrl++", triggered=self.largerFont))
+        menu2.addAction(QAction(config.thisTranslation["menu2_smaller"], self, shortcut="Ctrl+-", triggered=self.smallerFont))
+        menu2.addSeparator()
+        menu2.addAction(QAction(config.thisTranslation["menu2_format"], self, shortcut="Ctrl+P", triggered=self.enableParagraphButtonClicked))
+        menu2.addAction(QAction(config.thisTranslation["menu2_subHeadings"], self, triggered=self.enableSubheadingButtonClicked))
+        if config.enableVerseHighlighting:
+            menu2.addSeparator()
+            menu2.addAction(QAction(config.thisTranslation["menu2_toggleHighlightMarkers"], self, triggered=self.toggleHighlightMarker))
+
+        menu3 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu3_history"]))
+        menu3.addAction(QAction(config.thisTranslation["menu3_main"], self, shortcut="Ctrl+'", triggered=self.mainHistoryButtonClicked))
+        menu3.addAction(QAction(config.thisTranslation["menu3_mainBack"], self, shortcut="Ctrl+[", triggered=self.back))
+        menu3.addAction(QAction(config.thisTranslation["menu3_mainForward"], self, shortcut="Ctrl+]", triggered=self.forward))
+        menu3.addSeparator()
+        menu3.addAction(QAction(config.thisTranslation["menu3_study"], self, shortcut = 'Ctrl+"', triggered=self.studyHistoryButtonClicked))
+        menu3.addAction(QAction(config.thisTranslation["menu3_studyBack"], self, shortcut="Ctrl+{", triggered=self.studyBack))
+        menu3.addAction(QAction(config.thisTranslation["menu3_studyForward"], self, shortcut="Ctrl+}", triggered=self.studyForward))
+        menu3.addSeparator()
+        menu3.addAction(QAction(config.thisTranslation["menu1_reload"], self, triggered=self.reloadCurrentRecord))
+
+        menu4 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu4_further"]))
+        menu4.addAction(QAction(config.thisTranslation["menu4_book"], self, triggered=self.bookFeatures))
+        menu4.addAction(QAction(config.thisTranslation["menu4_chapter"], self, triggered=self.chapterFeatures))
+        menu4.addSeparator()
+        menu4.addAction(QAction(config.thisTranslation["menu4_next"], self, shortcut = 'Ctrl+>', triggered=self.nextMainChapter))
+        menu4.addAction(QAction(config.thisTranslation["menu4_previous"], self, shortcut = 'Ctrl+<', triggered=self.previousMainChapter))
+        menu4.addSeparator()
+        menu4.addAction(QAction(config.thisTranslation["menu4_indexes"], self, shortcut="Ctrl+.", triggered=self.runINDEX))
+        menu4.addAction(QAction(config.thisTranslation["menu4_commentary"], self, shortcut="Ctrl+Y", triggered=self.runCOMMENTARY))
+        menu4.addSeparator()
+        menu4.addAction(QAction(config.thisTranslation["menu4_traslations"], self, triggered=self.runTRANSLATION))
+        menu4.addAction(QAction(config.thisTranslation["menu4_discourse"], self, triggered=self.runDISCOURSE))
+        menu4.addAction(QAction(config.thisTranslation["menu4_words"], self, triggered=self.runWORDS))
+        menu4.addAction(QAction(config.thisTranslation["menu4_tdw"], self, shortcut="Ctrl+K", triggered=self.runCOMBO))
+        menu4.addSeparator()
+        menu4.addAction(QAction(config.thisTranslation["menu4_crossRef"], self, shortcut="Ctrl+R", triggered=self.runCROSSREFERENCE))
+        menu4.addAction(QAction(config.thisTranslation["menu4_tske"], self, shortcut="Ctrl+E", triggered=self.runTSKE))
+        menu4.addSeparator()
+        menu4.addAction(QAction(config.thisTranslation["menu4_compareAll"], self, shortcut="Ctrl+D", triggered=self.runCOMPARE))
+        menu4.addSeparator()
+        menu4.addAction(QAction(config.thisTranslation["menu4_moreComparison"], self, triggered=self.mainRefButtonClicked))
+
+        # check if books in favourite list exist
+        #for book in config.favouriteBooks:
+        #if not os.path.isfile(os.path.join(config.marvelData, "books", "{0}.book".format(book))):
+        #config.favouriteBooks.remove(book)
+
+        # remove an old book from favourite list if it is not installed
+        book = "Maps_ASB"
+        if not os.path.isfile(os.path.join(config.marvelData, "books", "{0}.book".format(book))) and book in config.favouriteBooks:
+            config.favouriteBooks.remove(book)
+
+        menu10 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu10_books"]))
+        if config.favouriteBooks:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[0]), self, triggered=self.openFavouriteBook0))
+        if len(config.favouriteBooks) > 1:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[1]), self, triggered=self.openFavouriteBook1))
+        if len(config.favouriteBooks) > 2:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[2]), self, triggered=self.openFavouriteBook2))
+        if len(config.favouriteBooks) > 3:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[3]), self, triggered=self.openFavouriteBook3))
+        if len(config.favouriteBooks) > 4:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[4]), self, triggered=self.openFavouriteBook4))
+        if len(config.favouriteBooks) > 5:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[5]), self, triggered=self.openFavouriteBook5))
+        if len(config.favouriteBooks) > 6:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[6]), self, triggered=self.openFavouriteBook6))
+        if len(config.favouriteBooks) > 7:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[7]), self, triggered=self.openFavouriteBook7))
+        if len(config.favouriteBooks) > 8:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[8]), self, triggered=self.openFavouriteBook8))
+        if len(config.favouriteBooks) > 9:
+            menu10.addAction(QAction(self.getBookName(config.favouriteBooks[9]), self, triggered=self.openFavouriteBook9))
+        menu10.addSeparator()
+        menu10.addAction(QAction(config.thisTranslation["menu5_book"], self, triggered=self.openBookMenu))
+        menu10.addAction(QAction(config.thisTranslation["menu10_dialog"], self, triggered=self.openBookDialog))
+        menu10.addSeparator()
+        menu10.addAction(QAction(config.thisTranslation["menu10_addFavourite"], self, triggered=self.addFavouriteBookDialog))
+        menu10.addSeparator()
+        menu10.addAction(QAction(config.thisTranslation["menu10_bookFromImages"], self, triggered=self.createBookModuleFromImages))
+        menu10.addAction(QAction(config.thisTranslation["menu10_bookFromHtml"], self, triggered=self.createBookModuleFromHTML))
+        menu10.addAction(QAction(config.thisTranslation["menu10_bookFromNotes"], self, triggered=self.createBookModuleFromNotes))
+        menu10.addSeparator()
+        menu10.addAction(QAction(config.thisTranslation["menu10_clearBookHighlights"], self, triggered=self.clearBookHighlights))
+
+        menu5 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu5_search"]))
+        menu5.addAction(QAction(config.thisTranslation["menu5_main"], self, shortcut="Ctrl+1", triggered=self.displaySearchBibleCommand))
+        menu5.addAction(QAction(config.thisTranslation["menu5_study"], self, shortcut="Ctrl+2", triggered=self.displaySearchStudyBibleCommand))
+        menu5.addAction(QAction(config.thisTranslation["menu5_bible"], self, triggered=self.displaySearchBibleMenu))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_dictionary"], self, shortcut="Ctrl+3", triggered=self.searchCommandBibleDictionary))
+        menu5.addAction(QAction(config.thisTranslation["context1_dict"], self, triggered=self.searchDictionaryDialog))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_encyclopedia"], self, shortcut="Ctrl+4", triggered=self.searchCommandBibleEncyclopedia))
+        menu5.addAction(QAction(config.thisTranslation["context1_encyclopedia"], self, triggered=self.searchEncyclopediaDialog))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_book"], self, shortcut="Ctrl+5", triggered=self.displaySearchBookCommand))
+        menu5.addAction(QAction(config.thisTranslation["menu5_selectBook"], self, triggered=self.searchBookDialog))
+        menu5.addAction(QAction(config.thisTranslation["menu5_favouriteBook"], self, triggered=self.displaySearchFavBookCommand))
+        menu5.addAction(QAction(config.thisTranslation["menu5_allBook"], self, triggered=self.displaySearchAllBookCommand))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_lastTopics"], self, shortcut="Ctrl+6", triggered=self.searchCommandBibleTopic))
+        menu5.addAction(QAction(config.thisTranslation["menu5_topics"], self, triggered=self.searchTopicDialog))
+        menu5.addAction(QAction(config.thisTranslation["menu5_allTopics"], self, triggered=self.searchCommandAllBibleTopic))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_characters"], self, shortcut="Ctrl+7", triggered=self.searchCommandBibleCharacter))
+        menu5.addAction(QAction(config.thisTranslation["menu5_names"], self, shortcut="Ctrl+8", triggered=self.searchCommandBibleName))
+        menu5.addAction(QAction(config.thisTranslation["menu5_locations"], self, shortcut="Ctrl+9", triggered=self.searchCommandBibleLocation))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_lexicon"], self, shortcut="Ctrl+0", triggered=self.searchCommandLexicon))
+        menu5.addSeparator()
+        menu5.addAction(QAction(config.thisTranslation["menu5_last3rdDict"], self, triggered=self.searchCommandThirdPartyDictionary))
+        menu5.addAction(QAction(config.thisTranslation["menu5_3rdDict"], self, triggered=self.search3rdDictionaryDialog))
+
+        menu6 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu6_notes"]))
+        menu6.addAction(QAction(config.thisTranslation["menu6_mainChapter"], self, triggered=self.openMainChapterNote))
+        menu6.addAction(QAction(config.thisTranslation["menu6_studyChapter"], self, triggered=self.openStudyChapterNote))
+        menu6.addAction(QAction(config.thisTranslation["menu6_searchChapters"], self, triggered=self.searchCommandChapterNote))
+        menu6.addSeparator()
+        menu6.addAction(QAction(config.thisTranslation["menu6_mainVerse"], self, triggered=self.openMainVerseNote))
+        menu6.addAction(QAction(config.thisTranslation["menu6_studyVerse"], self, triggered=self.openStudyVerseNote))
+        menu6.addAction(QAction(config.thisTranslation["menu6_searchVerses"], self, triggered=self.searchCommandVerseNote))
+        menu6.addSeparator()
+        menu6.addAction(QAction(config.thisTranslation["menu10_clearBookHighlights"], self, triggered=self.clearNoteHighlights))
+
+        menu7 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu7_topics"]))
+        menu7.addAction(QAction(config.thisTranslation["menu7_create"], self, shortcut="Ctrl+N", triggered=self.createNewNoteFile))
+        menu7.addAction(QAction(config.thisTranslation["menu7_open"], self, triggered=self.openTextFileDialog))
+        menu7.addSeparator()
+        menu7.addAction(QAction(config.thisTranslation["menu7_read"], self, triggered=self.externalFileButtonClicked))
+        menu7.addAction(QAction(config.thisTranslation["menu7_edit"], self, triggered=self.editExternalFileButtonClicked))
+        menu7.addSeparator()
+        menu7.addAction(QAction(config.thisTranslation["menu7_recent"], self, triggered=self.openExternalFileHistory))
+        menu7.addSeparator()
+        menu7.addAction(QAction(config.thisTranslation["menu7_paste"], self, shortcut="Ctrl+^", triggered=self.pasteFromClipboard))
+
+        menu11 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu11_multimedia"]))
+        menu11.addAction(QAction(config.thisTranslation["menu11_images"], self, triggered=self.openImagesFolder))
+        menu11.addAction(QAction(config.thisTranslation["menu11_music"], self, triggered=self.openMusicFolder))
+        menu11.addAction(QAction(config.thisTranslation["menu11_video"], self, triggered=self.openVideoFolder))
+        menu11.addSeparator()
+        menu11.addAction(QAction(config.thisTranslation["menu11_setupDownload"], self, triggered=self.setupYouTube))
+        menu11.addAction(QAction(config.thisTranslation["menu11_youtube"], self, triggered=self.openYouTube))
+        menu11.addSeparator()
+        menu11.addAction(QAction("YouTube -> mp3", self, triggered=self.downloadMp3Dialog))
+        menu11.addAction(QAction("YouTube -> mp4", self, triggered=self.downloadMp4Dialog))
+
+        menu8 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu8_resources"]))
+        menu8.addAction(QAction(config.thisTranslation["menu8_marvelData"], self, triggered=self.openMarvelDataFolder))
+        menu8.addSeparator()
+        menu8.addAction(QAction(config.thisTranslation["menu8_bibles"], self, triggered=self.installMarvelBibles))
+        menu8.addAction(QAction(config.thisTranslation["menu8_commentaries"], self, triggered=self.installMarvelCommentaries))
+        menu8.addAction(QAction(config.thisTranslation["menu8_datasets"], self, triggered=self.installMarvelDatasets))
+        menu8.addSeparator()
+        menu8.addAction(QAction(config.thisTranslation["menu8_plusLexicons"], self, triggered=self.importBBPlusLexiconInAFolder))
+        menu8.addAction(QAction(config.thisTranslation["menu8_plusDictionaries"], self, triggered=self.importBBPlusDictionaryInAFolder))
+        menu8.addSeparator()
+        menu8.addAction(QAction(config.thisTranslation["menu8_download3rdParty"], self, triggered=self.moreBooks))
+        menu8.addSeparator()
+        menu8.addAction(QAction(config.thisTranslation["menu8_3rdParty"], self, triggered=self.importModules))
+        menu8.addAction(QAction(config.thisTranslation["menu8_3rdPartyInFolder"], self, triggered=self.importModulesInFolder))
+        menu8.addAction(QAction(config.thisTranslation["menu8_settings"], self, triggered=self.importSettingsDialog))
+        menu8.addSeparator()
+        menu8.addAction(QAction(config.thisTranslation["menu8_tagFile"], self, triggered=self.tagFile))
+        menu8.addAction(QAction(config.thisTranslation["menu8_tagFiles"], self, triggered=self.tagFiles))
+        menu8.addAction(QAction(config.thisTranslation["menu8_tagFolder"], self, triggered=self.tagFolder))
+        menu8.addSeparator()
+        menu8.addAction(QAction(config.thisTranslation["menu8_fixDatabase"], self, triggered=self.selectDatabaseToFix))
+
+        if config.showInformation:
+            menu9 = self.menuBar().addMenu("&{0}".format(config.thisTranslation["menu9_information"]))
+            menu9.addAction(QAction(config.thisTranslation["menu1_wikiPages"], self, triggered=self.openUbaWiki))
+            menu9.addSeparator()
+            menu9.addAction(QAction("BibleTools.app", self, triggered=self.openBibleTools))
+            menu9.addAction(QAction("UniqueBible.app", self, triggered=self.openUniqueBible))
+            menu9.addAction(QAction("Marvel.bible", self, triggered=self.openMarvelBible))
+            menu9.addAction(QAction("BibleBento.com", self, triggered=self.openBibleBento))
+            menu9.addAction(QAction("OpenGNT.com", self, triggered=self.openOpenGNT))
+            menu9.addSeparator()
+            menu9.addAction(QAction("GitHub Repositories", self, triggered=self.openSource))
+            menu9.addAction(QAction("Unique Bible", self, triggered=self.openUniqueBibleSource))
+            menu9.addAction(QAction("Open Hebrew Bible", self, triggered=self.openHebrewBibleSource))
+            menu9.addAction(QAction("Open Greek New Testament", self, triggered=self.openOpenGNTSource))
+            menu9.addSeparator()
+            menu9.addAction(QAction("MarvelBible.Slack.com", self, triggered=self.goToSlack))
+            menu9.addAction(QAction(config.thisTranslation["menu9_contact"], self, triggered=self.contactEliranWong))
+            menu9.addSeparator()
+            menu9.addAction(QAction(config.thisTranslation["menu9_donate"], self, triggered=self.donateToUs))
+
+        if config.testing:
+            menu999 = self.menuBar().addMenu("&Testing")
+            menu999.addAction(QAction("Download Google Static Maps", self, triggered=self.downloadGoogleStaticMaps))
+
+    def setupToolBarStandardIconSize(self):
+
+        textButtonStyle = "QPushButton {background-color: #151B54; color: white;} QPushButton:hover {background-color: #333972;} QPushButton:pressed { background-color: #515790;}"
+
+        self.firstToolBar = QToolBar()
+        self.firstToolBar.setWindowTitle(config.thisTranslation["bar1_title"])
+        self.firstToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(self.firstToolBar)
+
+        self.mainTextMenuButton = QPushButton(self.verseReference("main")[0])
+        self.mainTextMenuButton.setToolTip(config.thisTranslation["bar1_menu"])
+        self.mainTextMenuButton.setStyleSheet(textButtonStyle)
+        self.mainTextMenuButton.clicked.connect(self.mainTextMenu)
+        self.firstToolBar.addWidget(self.mainTextMenuButton)
+
+        self.mainRefButton = QPushButton(self.verseReference("main")[1])
+        self.mainRefButton.setToolTip(config.thisTranslation["bar1_reference"])
+        self.mainRefButton.setStyleSheet(textButtonStyle)
+        self.mainRefButton.clicked.connect(self.mainRefButtonClicked)
+        self.firstToolBar.addWidget(self.mainRefButton)
+
+        openChapterNoteButton = QPushButton()
+        #openChapterNoteButton.setFixedSize(40, 40)
+        #openChapterNoteButton.setBaseSize(70, 70)
+        openChapterNoteButton.setToolTip(config.thisTranslation["bar1_chapterNotes"])
+        openChapterNoteButtonFile = os.path.join("htmlResources", "noteChapter.png")
+        openChapterNoteButton.setIcon(QIcon(openChapterNoteButtonFile))
+        openChapterNoteButton.clicked.connect(self.openMainChapterNote)
+        self.firstToolBar.addWidget(openChapterNoteButton)
+        #t = self.firstToolBar.addAction(QIcon(openChapterNoteButtonFile), "Main View Chapter Notes", self.openMainChapterNote)
+        #openChapterNoteButtonFile = os.path.join("htmlResources", "search.png")
+        #t.setIcon(QIcon(openChapterNoteButtonFile))
+
+        openVerseNoteButton = QPushButton()
+        openVerseNoteButton.setToolTip(config.thisTranslation["bar1_verseNotes"])
+        openVerseNoteButtonFile = os.path.join("htmlResources", "noteVerse.png")
+        openVerseNoteButton.setIcon(QIcon(openVerseNoteButtonFile))
+        openVerseNoteButton.clicked.connect(self.openMainVerseNote)
+        self.firstToolBar.addWidget(openVerseNoteButton)
+
+        searchBibleButton = QPushButton()
+        searchBibleButton.setToolTip(config.thisTranslation["bar1_searchBible"])
+        searchBibleButtonFile = os.path.join("htmlResources", "search.png")
+        searchBibleButton.setIcon(QIcon(searchBibleButtonFile))
+        searchBibleButton.clicked.connect(self.displaySearchBibleCommand)
+        self.firstToolBar.addWidget(searchBibleButton)
+
+        searchBibleButton = QPushButton()
+        searchBibleButton.setToolTip(config.thisTranslation["bar1_searchBibles"])
+        searchBibleButtonFile = os.path.join("htmlResources", "search_plus.png")
+        searchBibleButton.setIcon(QIcon(searchBibleButtonFile))
+        searchBibleButton.clicked.connect(self.displaySearchBibleMenu)
+        self.firstToolBar.addWidget(searchBibleButton)
+
+        self.firstToolBar.addSeparator()
+
+        self.textCommandLineEdit = QLineEdit()
+        self.textCommandLineEdit.setToolTip(config.thisTranslation["bar1_command"])
+        self.textCommandLineEdit.setMinimumWidth(100)
+        self.textCommandLineEdit.returnPressed.connect(self.textCommandEntered)
+        self.firstToolBar.addWidget(self.textCommandLineEdit)
+
+        self.firstToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["bar1_toolbars"])
+        actionButtonFile = os.path.join("htmlResources", "toolbar.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.hideShowAdditionalToolBar)
+        self.firstToolBar.addWidget(actionButton)
+
+        self.addToolBarBreak()
+
+        self.studyBibleToolBar = QToolBar()
+        self.studyBibleToolBar.setWindowTitle(config.thisTranslation["bar2_title"])
+        self.studyBibleToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(self.studyBibleToolBar)
+
+        self.studyTextMenuButton = QPushButton(self.verseReference("study")[0])
+        self.studyTextMenuButton.setToolTip(config.thisTranslation["bar2_menu"])
+        self.studyTextMenuButton.setStyleSheet(textButtonStyle)
+        self.studyTextMenuButton.clicked.connect(self.studyTextMenu)
+        self.studyBibleToolBar.addWidget(self.studyTextMenuButton)
+
+        self.studyRefButton = QPushButton(self.verseReference("study")[1])
+        self.studyRefButton.setToolTip(config.thisTranslation["bar2_reference"])
+        self.studyRefButton.setStyleSheet(textButtonStyle)
+        self.studyRefButton.clicked.connect(self.studyRefButtonClicked)
+        self.studyBibleToolBar.addWidget(self.studyRefButton)
+
+        openStudyChapterNoteButton = QPushButton()
+        openStudyChapterNoteButton.setToolTip(config.thisTranslation["bar2_chapterNotes"])
+        openStudyChapterNoteButtonFile = os.path.join("htmlResources", "noteChapter.png")
+        openStudyChapterNoteButton.setIcon(QIcon(openStudyChapterNoteButtonFile))
+        openStudyChapterNoteButton.clicked.connect(self.openStudyChapterNote)
+        self.studyBibleToolBar.addWidget(openStudyChapterNoteButton)
+
+        openStudyVerseNoteButton = QPushButton()
+        openStudyVerseNoteButton.setToolTip(config.thisTranslation["bar2_verseNotes"])
+        openStudyVerseNoteButtonFile = os.path.join("htmlResources", "noteVerse.png")
+        openStudyVerseNoteButton.setIcon(QIcon(openStudyVerseNoteButtonFile))
+        openStudyVerseNoteButton.clicked.connect(self.openStudyVerseNote)
+        self.studyBibleToolBar.addWidget(openStudyVerseNoteButton)
+
+        searchStudyBibleButton = QPushButton()
+        searchStudyBibleButton.setToolTip(config.thisTranslation["bar2_searchBible"])
+        searchStudyBibleButtonFile = os.path.join("htmlResources", "search.png")
+        searchStudyBibleButton.setIcon(QIcon(searchStudyBibleButtonFile))
+        searchStudyBibleButton.clicked.connect(self.displaySearchStudyBibleCommand)
+        self.studyBibleToolBar.addWidget(searchStudyBibleButton)
+
+        searchStudyBibleButton = QPushButton()
+        searchStudyBibleButton.setToolTip(config.thisTranslation["bar2_searchBibles"])
+        searchStudyBibleButtonFile = os.path.join("htmlResources", "search_plus.png")
+        searchStudyBibleButton.setIcon(QIcon(searchStudyBibleButtonFile))
+        searchStudyBibleButton.clicked.connect(self.displaySearchBibleMenu)
+        self.studyBibleToolBar.addWidget(searchStudyBibleButton)
+
+        self.enableSyncStudyWindowBibleButton = QPushButton()
+        self.enableSyncStudyWindowBibleButton.setToolTip(self.getSyncStudyWindowBibleDisplayToolTip())
+        enableSyncStudyWindowBibleButtonFile = os.path.join("htmlResources", self.getSyncStudyWindowBibleDisplay())
+        self.enableSyncStudyWindowBibleButton.setIcon(QIcon(enableSyncStudyWindowBibleButtonFile))
+        self.enableSyncStudyWindowBibleButton.clicked.connect(self.enableSyncStudyWindowBibleButtonClicked)
+        self.studyBibleToolBar.addWidget(self.enableSyncStudyWindowBibleButton)
+
+        self.secondToolBar = QToolBar()
+        self.secondToolBar.setWindowTitle(config.thisTranslation["bar2_title"])
+        self.secondToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(self.secondToolBar)
+
+        self.enableStudyBibleButton = QPushButton()
+        self.enableStudyBibleButton.setToolTip(self.getStudyBibleDisplayToolTip())
+        enableStudyBibleButtonFile = os.path.join("htmlResources", self.getStudyBibleDisplay())
+        self.enableStudyBibleButton.setIcon(QIcon(enableStudyBibleButtonFile))
+        self.enableStudyBibleButton.clicked.connect(self.enableStudyBibleButtonClicked)
+        self.secondToolBar.addWidget(self.enableStudyBibleButton)
+
+        self.secondToolBar.addSeparator()
+
+        self.commentaryRefButton = QPushButton(self.verseReference("commentary"))
+        self.commentaryRefButton.setToolTip(config.thisTranslation["menu4_commentary"])
+        self.commentaryRefButton.setStyleSheet(textButtonStyle)
+        self.commentaryRefButton.clicked.connect(self.commentaryRefButtonClicked)
+        self.secondToolBar.addWidget(self.commentaryRefButton)
+
+        self.enableSyncCommentaryButton = QPushButton()
+        self.enableSyncCommentaryButton.setToolTip(self.getSyncCommentaryDisplayToolTip())
+        enableSyncCommentaryButtonFile = os.path.join("htmlResources", self.getSyncCommentaryDisplay())
+        self.enableSyncCommentaryButton.setIcon(QIcon(enableSyncCommentaryButtonFile))
+        self.enableSyncCommentaryButton.clicked.connect(self.enableSyncCommentaryButtonClicked)
+        self.secondToolBar.addWidget(self.enableSyncCommentaryButton)
+
+        self.secondToolBar.addSeparator()
+
+        openFileButton = QPushButton()
+        openFileButton.setToolTip(config.thisTranslation["menu10_dialog"])
+        openFileButtonFile = os.path.join("htmlResources", "open.png")
+        openFileButton.setIcon(QIcon(openFileButtonFile))
+        openFileButton.clicked.connect(self.openBookDialog)
+        self.secondToolBar.addWidget(openFileButton)
+
+        self.bookButton = QPushButton(config.book)
+        self.bookButton.setToolTip(config.thisTranslation["menu5_book"])
+        self.bookButton.setStyleSheet(textButtonStyle)
+        self.bookButton.clicked.connect(self.openBookMenu)
+        self.secondToolBar.addWidget(self.bookButton)
+
+        searchBookButton = QPushButton()
+        searchBookButton.setToolTip(config.thisTranslation["bar2_searchBooks"])
+        searchBookButtonFile = os.path.join("htmlResources", "search.png")
+        searchBookButton.setIcon(QIcon(searchBookButtonFile))
+        searchBookButton.clicked.connect(self.displaySearchBookCommand)
+        self.secondToolBar.addWidget(searchBookButton)
+
+        self.secondToolBar.addSeparator()
+
+        newFileButton = QPushButton()
+        newFileButton.setToolTip(config.thisTranslation["menu7_create"])
+        newFileButtonFile = os.path.join("htmlResources", "newfile.png")
+        newFileButton.setIcon(QIcon(newFileButtonFile))
+        newFileButton.clicked.connect(self.createNewNoteFile)
+        self.secondToolBar.addWidget(newFileButton)
+
+        openFileButton = QPushButton()
+        openFileButton.setToolTip(config.thisTranslation["menu7_open"])
+        openFileButtonFile = os.path.join("htmlResources", "open.png")
+        openFileButton.setIcon(QIcon(openFileButtonFile))
+        openFileButton.clicked.connect(self.openTextFileDialog)
+        self.secondToolBar.addWidget(openFileButton)
+
+        self.externalFileButton = QPushButton(self.getLastExternalFileName())
+        self.externalFileButton.setToolTip(config.thisTranslation["menu7_read"])
+        self.externalFileButton.setStyleSheet(textButtonStyle)
+        self.externalFileButton.clicked.connect(self.externalFileButtonClicked)
+        self.secondToolBar.addWidget(self.externalFileButton)
+
+        editExternalFileButton = QPushButton()
+        editExternalFileButton.setToolTip(config.thisTranslation["menu7_edit"])
+        editExternalFileButtonFile = os.path.join("htmlResources", "edit.png")
+        editExternalFileButton.setIcon(QIcon(editExternalFileButtonFile))
+        editExternalFileButton.clicked.connect(self.editExternalFileButtonClicked)
+        self.secondToolBar.addWidget(editExternalFileButton)
+
+        self.secondToolBar.addSeparator()
+
+        fontMinusButton = QPushButton()
+        fontMinusButton.setToolTip(config.thisTranslation["menu2_smaller"])
+        fontMinusButtonFile = os.path.join("htmlResources", "fontMinus.png")
+        fontMinusButton.setIcon(QIcon(fontMinusButtonFile))
+        fontMinusButton.clicked.connect(self.smallerFont)
+        self.secondToolBar.addWidget(fontMinusButton)
+
+        self.defaultFontButton = QPushButton("{0} {1}".format(config.font, config.fontSize))
+        self.defaultFontButton.setToolTip(config.thisTranslation["menu1_setDefaultFont"])
+        self.defaultFontButton.setStyleSheet(textButtonStyle)
+        self.defaultFontButton.clicked.connect(self.setDefaultFont)
+        self.secondToolBar.addWidget(self.defaultFontButton)
+
+        fontPlusButton = QPushButton()
+        fontPlusButton.setToolTip(config.thisTranslation["menu2_larger"])
+        fontPlusButtonFile = os.path.join("htmlResources", "fontPlus.png")
+        fontPlusButton.setIcon(QIcon(fontPlusButtonFile))
+        fontPlusButton.clicked.connect(self.largerFont)
+        self.secondToolBar.addWidget(fontPlusButton)
+
+        self.secondToolBar.addSeparator()
+
+        youtubeButton = QPushButton()
+        youtubeButton.setToolTip(config.thisTranslation["menu11_youtube"])
+        youtubeButtonFile = os.path.join("htmlResources", "youtube.png")
+        youtubeButton.setIcon(QIcon(youtubeButtonFile))
+        youtubeButton.clicked.connect(self.openYouTube)
+        self.secondToolBar.addWidget(youtubeButton)
+
+        self.secondToolBar.addSeparator()
+
+        reloadButton = QPushButton()
+        reloadButton.setToolTip(config.thisTranslation["menu1_reload"])
+        reloadButtonFile = os.path.join("htmlResources", "reload.png")
+        reloadButton.setIcon(QIcon(reloadButtonFile))
+        reloadButton.clicked.connect(self.reloadCurrentRecord)
+        self.secondToolBar.addWidget(reloadButton)
+
+        self.secondToolBar.addSeparator()
+
+        self.leftToolBar = QToolBar()
+        self.leftToolBar.setWindowTitle(config.thisTranslation["bar3_title"])
+        self.leftToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(Qt.LeftToolBarArea, self.leftToolBar)
+
+        backButton = QPushButton()
+        backButton.setToolTip(config.thisTranslation["menu3_mainBack"])
+        leftButtonFile = os.path.join("htmlResources", "left.png")
+        backButton.setIcon(QIcon(leftButtonFile))
+        backButton.clicked.connect(self.back)
+        self.leftToolBar.addWidget(backButton)
+
+        mainHistoryButton = QPushButton()
+        mainHistoryButton.setToolTip(config.thisTranslation["menu3_main"])
+        mainHistoryButtonFile = os.path.join("htmlResources", "history.png")
+        mainHistoryButton.setIcon(QIcon(mainHistoryButtonFile))
+        mainHistoryButton.clicked.connect(self.mainHistoryButtonClicked)
+        self.leftToolBar.addWidget(mainHistoryButton)
+
+        forwardButton = QPushButton()
+        forwardButton.setToolTip(config.thisTranslation["menu3_mainForward"])
+        rightButtonFile = os.path.join("htmlResources", "right.png")
+        forwardButton.setIcon(QIcon(rightButtonFile))
+        forwardButton.clicked.connect(self.forward)
+        self.leftToolBar.addWidget(forwardButton)
+
+        self.leftToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["bar3_pdf"])
+        actionButtonFile = os.path.join("htmlResources", "pdf.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.printMainPage)
+        self.leftToolBar.addWidget(actionButton)
+
+        self.leftToolBar.addSeparator()
+
+        self.enableParagraphButton = QPushButton()
+        self.enableParagraphButton.setToolTip(config.thisTranslation["menu2_format"])
+        enableParagraphButtonFile = os.path.join("htmlResources", self.getReadFormattedBibles())
+        self.enableParagraphButton.setIcon(QIcon(enableParagraphButtonFile))
+        self.enableParagraphButton.clicked.connect(self.enableParagraphButtonClicked)
+        self.leftToolBar.addWidget(self.enableParagraphButton)
+
+        self.enableSubheadingButton = QPushButton()
+        self.enableSubheadingButton.setToolTip(config.thisTranslation["menu2_subHeadings"])
+        enableSubheadingButtonFile = os.path.join("htmlResources", self.getAddSubheading())
+        self.enableSubheadingButton.setIcon(QIcon(enableSubheadingButtonFile))
+        self.enableSubheadingButton.clicked.connect(self.enableSubheadingButtonClicked)
+        self.leftToolBar.addWidget(self.enableSubheadingButton)
+
+        self.leftToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_previous"])
+        actionButtonFile = os.path.join("htmlResources", "previousChapter.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.previousMainChapter)
+        self.leftToolBar.addWidget(actionButton)
+
+        #        actionButton = QPushButton()
+        #        actionButton.setToolTip(config.thisTranslation["bar1_reference"])
+        #        actionButtonFile = os.path.join("htmlResources", "bible.png")
+        #        actionButton.setIcon(QIcon(actionButtonFile))
+        #        actionButton.clicked.connect(self.openMainChapter)
+        #        self.leftToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_next"])
+        actionButtonFile = os.path.join("htmlResources", "nextChapter.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.nextMainChapter)
+        self.leftToolBar.addWidget(actionButton)
+
+        self.leftToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_compareAll"])
+        actionButtonFile = os.path.join("htmlResources", "compare_with.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runCOMPARE)
+        self.leftToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_moreComparison"])
+        actionButtonFile = os.path.join("htmlResources", "parallel_with.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.mainRefButtonClicked)
+        self.leftToolBar.addWidget(actionButton)
+
+        self.leftToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip("Marvel Original Bible")
+        actionButtonFile = os.path.join("htmlResources", "original.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runMOB)
+        self.leftToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip("Marvel Interlinear Bible")
+        actionButtonFile = os.path.join("htmlResources", "interlinear.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runMIB)
+        self.leftToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip("Marvel Trilingual Bible")
+        actionButtonFile = os.path.join("htmlResources", "trilingual.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runMTB)
+        self.leftToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip("Marvel Parallel Bible")
+        actionButtonFile = os.path.join("htmlResources", "line.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runMPB)
+        self.leftToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip("Marvel Annotated Bible")
+        actionButtonFile = os.path.join("htmlResources", "annotated.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runMAB)
+        self.leftToolBar.addWidget(actionButton)
+
+        self.leftToolBar.addSeparator()
+
+        self.rightToolBar = QToolBar()
+        self.rightToolBar.setWindowTitle(config.thisTranslation["bar4_title"])
+        self.rightToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(Qt.RightToolBarArea, self.rightToolBar)
+
+        studyBackButton = QPushButton()
+        studyBackButton.setToolTip(config.thisTranslation["menu3_studyBack"])
+        leftButtonFile = os.path.join("htmlResources", "left.png")
+        studyBackButton.setIcon(QIcon(leftButtonFile))
+        studyBackButton.clicked.connect(self.studyBack)
+        self.rightToolBar.addWidget(studyBackButton)
+
+        studyHistoryButton = QPushButton()
+        studyHistoryButton.setToolTip(config.thisTranslation["menu3_study"])
+        studyHistoryButtonFile = os.path.join("htmlResources", "history.png")
+        studyHistoryButton.setIcon(QIcon(studyHistoryButtonFile))
+        studyHistoryButton.clicked.connect(self.studyHistoryButtonClicked)
+        self.rightToolBar.addWidget(studyHistoryButton)
+
+        studyForwardButton = QPushButton()
+        studyForwardButton.setToolTip(config.thisTranslation["menu3_studyForward"])
+        rightButtonFile = os.path.join("htmlResources", "right.png")
+        studyForwardButton.setIcon(QIcon(rightButtonFile))
+        studyForwardButton.clicked.connect(self.studyForward)
+        self.rightToolBar.addWidget(studyForwardButton)
+
+        self.rightToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["bar3_pdf"])
+        actionButtonFile = os.path.join("htmlResources", "pdf.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.printStudyPage)
+        self.rightToolBar.addWidget(actionButton)
+
+        self.rightToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_indexes"])
+        actionButtonFile = os.path.join("htmlResources", "indexes.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runINDEX)
+        self.rightToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_commentary"])
+        actionButtonFile = os.path.join("htmlResources", "commentary.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runCOMMENTARY)
+        self.rightToolBar.addWidget(actionButton)
+
+        self.rightToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_crossRef"])
+        actionButtonFile = os.path.join("htmlResources", "cross_reference.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runCROSSREFERENCE)
+        self.rightToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_tske"])
+        actionButtonFile = os.path.join("htmlResources", "treasure.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runTSKE)
+        self.rightToolBar.addWidget(actionButton)
+
+        self.rightToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_traslations"])
+        actionButtonFile = os.path.join("htmlResources", "translations.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runTRANSLATION)
+        self.rightToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_discourse"])
+        actionButtonFile = os.path.join("htmlResources", "discourse.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runDISCOURSE)
+        self.rightToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_words"])
+        actionButtonFile = os.path.join("htmlResources", "words.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runWORDS)
+        self.rightToolBar.addWidget(actionButton)
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu4_tdw"])
+        actionButtonFile = os.path.join("htmlResources", "combo.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.runCOMBO)
+        self.rightToolBar.addWidget(actionButton)
+
+        self.rightToolBar.addSeparator()
+
+        actionButton = QPushButton()
+        actionButton.setToolTip(config.thisTranslation["menu2_landscape"])
+        actionButtonFile = os.path.join("htmlResources", "portrait.png")
+        actionButton.setIcon(QIcon(actionButtonFile))
+        actionButton.clicked.connect(self.switchLandscapeMode)
+        self.rightToolBar.addWidget(actionButton)
+
+        parallelButton = QPushButton()
+        parallelButton.setToolTip(config.thisTranslation["menu2_study"])
+        parallelButtonFile = os.path.join("htmlResources", "parallel.png")
+        parallelButton.setIcon(QIcon(parallelButtonFile))
+        parallelButton.clicked.connect(self.parallel)
+        self.rightToolBar.addWidget(parallelButton)
+
+        self.rightToolBar.addSeparator()
+
+        self.enableInstantButton = QPushButton()
+        self.enableInstantButton.setToolTip(config.thisTranslation["menu2_hover"])
+        enableInstantButtonFile = os.path.join("htmlResources", self.getInstantInformation())
+        self.enableInstantButton.setIcon(QIcon(enableInstantButtonFile))
+        self.enableInstantButton.clicked.connect(self.enableInstantButtonClicked)
+        self.rightToolBar.addWidget(self.enableInstantButton)
+
+        instantButton = QPushButton()
+        instantButton.setToolTip(config.thisTranslation["menu2_bottom"])
+        instantButtonFile = os.path.join("htmlResources", "lightning.png")
+        instantButton.setIcon(QIcon(instantButtonFile))
+        instantButton.clicked.connect(self.cycleInstant)
+        self.rightToolBar.addWidget(instantButton)
+
+        self.rightToolBar.addSeparator()
+
+    def setupToolBarFullIconSize(self):
+
+        textButtonStyle = "QPushButton {background-color: #151B54; color: white;} QPushButton:hover {background-color: #333972;} QPushButton:pressed { background-color: #515790;}"
+
+        self.firstToolBar = QToolBar()
+        self.firstToolBar.setWindowTitle(config.thisTranslation["bar1_title"])
+        self.firstToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(self.firstToolBar)
+
+        self.mainTextMenuButton = QPushButton(self.verseReference("main")[0])
+        self.mainTextMenuButton.setToolTip(config.thisTranslation["bar1_menu"])
+        self.mainTextMenuButton.setStyleSheet(textButtonStyle)
+        self.mainTextMenuButton.clicked.connect(self.mainTextMenu)
+        self.firstToolBar.addWidget(self.mainTextMenuButton)
+
+        self.mainRefButton = QPushButton(self.verseReference("main")[1])
+        self.mainRefButton.setToolTip(config.thisTranslation["bar1_reference"])
+        self.mainRefButton.setStyleSheet(textButtonStyle)
+        self.mainRefButton.clicked.connect(self.mainRefButtonClicked)
+        self.firstToolBar.addWidget(self.mainRefButton)
+
+        iconFile = os.path.join("htmlResources", "noteChapter.png")
+        self.firstToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar1_chapterNotes"], self.openMainChapterNote)
+
+        iconFile = os.path.join("htmlResources", "noteVerse.png")
+        self.firstToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar1_verseNotes"], self.openMainVerseNote)
+
+        iconFile = os.path.join("htmlResources", "search.png")
+        self.firstToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar1_searchBible"], self.displaySearchBibleCommand)
+
+        iconFile = os.path.join("htmlResources", "search_plus.png")
+        self.firstToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar1_searchBibles"], self.displaySearchBibleMenu)
+
+        self.firstToolBar.addSeparator()
+
+        self.textCommandLineEdit = QLineEdit()
+        self.textCommandLineEdit.setToolTip(config.thisTranslation["bar1_command"])
+        self.textCommandLineEdit.setMinimumWidth(100)
+        self.textCommandLineEdit.returnPressed.connect(self.textCommandEntered)
+        self.firstToolBar.addWidget(self.textCommandLineEdit)
+
+        self.firstToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "toolbar.png")
+        self.firstToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar1_toolbars"], self.hideShowAdditionalToolBar)
+
+        self.addToolBarBreak()
+
+        self.studyBibleToolBar = QToolBar()
+        self.studyBibleToolBar.setWindowTitle(config.thisTranslation["bar2_title"])
+        self.studyBibleToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(self.studyBibleToolBar)
+
+        self.studyTextMenuButton = QPushButton(self.verseReference("study")[0])
+        self.studyTextMenuButton.setToolTip(config.thisTranslation["bar2_menu"])
+        self.studyTextMenuButton.setStyleSheet(textButtonStyle)
+        self.studyTextMenuButton.clicked.connect(self.studyTextMenu)
+        self.studyBibleToolBar.addWidget(self.studyTextMenuButton)
+
+        self.studyRefButton = QPushButton(self.verseReference("study")[1])
+        self.studyRefButton.setToolTip(config.thisTranslation["bar2_reference"])
+        self.studyRefButton.setStyleSheet(textButtonStyle)
+        self.studyRefButton.clicked.connect(self.studyRefButtonClicked)
+        self.studyBibleToolBar.addWidget(self.studyRefButton)
+
+        iconFile = os.path.join("htmlResources", "noteChapter.png")
+        self.studyBibleToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar2_chapterNotes"], self.openStudyChapterNote)
+
+        iconFile = os.path.join("htmlResources", "noteVerse.png")
+        self.studyBibleToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar2_verseNotes"], self.openStudyVerseNote)
+
+        iconFile = os.path.join("htmlResources", "search.png")
+        self.studyBibleToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar2_searchBible"], self.displaySearchStudyBibleCommand)
+
+        iconFile = os.path.join("htmlResources", "search_plus.png")
+        self.studyBibleToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar2_searchBibles"], self.displaySearchBibleMenu)
+
+        iconFile = os.path.join("htmlResources", self.getSyncStudyWindowBibleDisplay())
+        self.enableSyncStudyWindowBibleButton = self.studyBibleToolBar.addAction(QIcon(iconFile), self.getSyncStudyWindowBibleDisplayToolTip(), self.enableSyncStudyWindowBibleButtonClicked)
+
+        self.secondToolBar = QToolBar()
+        self.secondToolBar.setWindowTitle(config.thisTranslation["bar2_title"])
+        self.secondToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(self.secondToolBar)
+
+        iconFile = os.path.join("htmlResources", self.getStudyBibleDisplay())
+        self.enableStudyBibleButton = self.secondToolBar.addAction(QIcon(iconFile), self.getStudyBibleDisplayToolTip(), self.enableStudyBibleButtonClicked)
+
+        self.secondToolBar.addSeparator()
+
+        self.commentaryRefButton = QPushButton(self.verseReference("commentary"))
+        self.commentaryRefButton.setToolTip(config.thisTranslation["menu4_commentary"])
+        self.commentaryRefButton.setStyleSheet(textButtonStyle)
+        self.commentaryRefButton.clicked.connect(self.commentaryRefButtonClicked)
+        self.secondToolBar.addWidget(self.commentaryRefButton)
+
+        iconFile = os.path.join("htmlResources", self.getSyncCommentaryDisplay())
+        self.enableSyncCommentaryButton = self.secondToolBar.addAction(QIcon(iconFile), self.getSyncCommentaryDisplayToolTip(), self.enableSyncCommentaryButtonClicked)
+
+        self.secondToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "open.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu10_dialog"], self.openBookDialog)
+
+        self.bookButton = QPushButton(config.book)
+        self.bookButton.setToolTip(config.thisTranslation["menu5_book"])
+        self.bookButton.setStyleSheet(textButtonStyle)
+        self.bookButton.clicked.connect(self.openBookMenu)
+        self.secondToolBar.addWidget(self.bookButton)
+
+        iconFile = os.path.join("htmlResources", "search.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar2_searchBooks"], self.displaySearchBookCommand)
+
+        self.secondToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "newfile.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu7_create"], self.createNewNoteFile)
+
+        iconFile = os.path.join("htmlResources", "open.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu7_open"], self.openTextFileDialog)
+
+        self.externalFileButton = QPushButton(self.getLastExternalFileName())
+        self.externalFileButton.setToolTip(config.thisTranslation["menu7_read"])
+        self.externalFileButton.setStyleSheet(textButtonStyle)
+        self.externalFileButton.clicked.connect(self.externalFileButtonClicked)
+        self.secondToolBar.addWidget(self.externalFileButton)
+
+        iconFile = os.path.join("htmlResources", "edit.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu7_edit"], self.editExternalFileButtonClicked)
+
+        self.secondToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "fontMinus.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_smaller"], self.smallerFont)
+
+        self.defaultFontButton = QPushButton("{0} {1}".format(config.font, config.fontSize))
+        self.defaultFontButton.setToolTip(config.thisTranslation["menu1_setDefaultFont"])
+        self.defaultFontButton.setStyleSheet(textButtonStyle)
+        self.defaultFontButton.clicked.connect(self.setDefaultFont)
+        self.secondToolBar.addWidget(self.defaultFontButton)
+
+        iconFile = os.path.join("htmlResources", "fontPlus.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_larger"], self.largerFont)
+
+        self.secondToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "youtube.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu11_youtube"], self.openYouTube)
+
+        self.secondToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "reload.png")
+        self.secondToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu1_reload"], self.reloadCurrentRecord)
+
+        self.secondToolBar.addSeparator()
+
+        self.leftToolBar = QToolBar()
+        self.leftToolBar.setWindowTitle(config.thisTranslation["bar3_title"])
+        self.leftToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(Qt.LeftToolBarArea, self.leftToolBar)
+
+        iconFile = os.path.join("htmlResources", "left.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu3_mainBack"], self.back)
+
+        iconFile = os.path.join("htmlResources", "history.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu3_main"], self.mainHistoryButtonClicked)
+
+        iconFile = os.path.join("htmlResources", "right.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu3_mainForward"], self.forward)
+
+        self.leftToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "pdf.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar3_pdf"], self.printMainPage)
+
+        self.leftToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", self.getReadFormattedBibles())
+        self.enableParagraphButton = self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_format"], self.enableParagraphButtonClicked)
+
+        iconFile = os.path.join("htmlResources", self.getAddSubheading())
+        self.enableSubheadingButton = self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_subHeadings"], self.enableSubheadingButtonClicked)
+
+        self.leftToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "previousChapter.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_previous"], self.previousMainChapter)
+
+        #        actionButton = QPushButton()
+        #        actionButton.setToolTip(config.thisTranslation["bar1_reference"])
+        #        actionButtonFile = os.path.join("htmlResources", "bible.png")
+        #        actionButton.setIcon(QIcon(actionButtonFile))
+        #        actionButton.clicked.connect(self.openMainChapter)
+        #        self.leftToolBar.addWidget(actionButton)
+
+        iconFile = os.path.join("htmlResources", "nextChapter.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_next"], self.nextMainChapter)
+
+        self.leftToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "compare_with.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_compareAll"], self.runCOMPARE)
+
+        iconFile = os.path.join("htmlResources", "parallel_with.png")
+        self.leftToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_moreComparison"], self.mainRefButtonClicked)
+
+        self.leftToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "original.png")
+        self.leftToolBar.addAction(QIcon(iconFile), "Marvel Original Bible", self.runMOB)
+
+        iconFile = os.path.join("htmlResources", "interlinear.png")
+        self.leftToolBar.addAction(QIcon(iconFile), "Marvel Interlinear Bible", self.runMIB)
+
+        iconFile = os.path.join("htmlResources", "trilingual.png")
+        self.leftToolBar.addAction(QIcon(iconFile), "Marvel Trilingual Bible", self.runMTB)
+
+        iconFile = os.path.join("htmlResources", "line.png")
+        self.leftToolBar.addAction(QIcon(iconFile), "Marvel Parallel Bible", self.runMPB)
+
+        iconFile = os.path.join("htmlResources", "annotated.png")
+        self.leftToolBar.addAction(QIcon(iconFile), "Marvel Annotated Bible", self.runMAB)
+
+        self.leftToolBar.addSeparator()
+
+        self.rightToolBar = QToolBar()
+        self.rightToolBar.setWindowTitle(config.thisTranslation["bar4_title"])
+        self.rightToolBar.setContextMenuPolicy(Qt.PreventContextMenu)
+        self.addToolBar(Qt.RightToolBarArea, self.rightToolBar)
+
+        iconFile = os.path.join("htmlResources", "left.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu3_studyBack"], self.studyBack)
+
+        iconFile = os.path.join("htmlResources", "history.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu3_study"], self.studyHistoryButtonClicked)
+
+        iconFile = os.path.join("htmlResources", "right.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu3_studyForward"], self.studyForward)
+
+        self.rightToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "pdf.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["bar3_pdf"], self.printStudyPage)
+
+        self.rightToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "indexes.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_indexes"], self.runINDEX)
+
+        iconFile = os.path.join("htmlResources", "commentary.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_commentary"], self.runCOMMENTARY)
+
+        self.rightToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "cross_reference.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_crossRef"], self.runCROSSREFERENCE)
+
+        iconFile = os.path.join("htmlResources", "treasure.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_tske"], self.runTSKE)
+
+        self.rightToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "translations.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_traslations"], self.runTRANSLATION)
+
+        iconFile = os.path.join("htmlResources", "discourse.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_discourse"], self.runDISCOURSE)
+
+        iconFile = os.path.join("htmlResources", "words.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_words"], self.runWORDS)
+
+        iconFile = os.path.join("htmlResources", "combo.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu4_tdw"], self.runCOMBO)
+
+        self.rightToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", "portrait.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_landscape"], self.switchLandscapeMode)
+
+        iconFile = os.path.join("htmlResources", "parallel.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_study"], self.parallel)
+
+        self.rightToolBar.addSeparator()
+
+        iconFile = os.path.join("htmlResources", self.getInstantInformation())
+        self.enableInstantButton = self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_hover"], self.enableInstantButtonClicked)
+
+        iconFile = os.path.join("htmlResources", "lightning.png")
+        self.rightToolBar.addAction(QIcon(iconFile), config.thisTranslation["menu2_bottom"], self.cycleInstant)
+
+        self.rightToolBar.addSeparator()
