+++ latest_changes.txt
@@ -1,3 +1,12 @@
+Latest change in version 19.5:
+Merged Oliver's new work on verse highlighting:
+- Verse highlighting released for testing. Enable it by checking enableVerseHighlighting and restarting app
+- Toggle the markers on and off in Layout / Toggle Highlight Markers
+- Highlight markers show up only in verse view to the left of the verses
+- There are currently 3 markers (yellow, blue, and underline)
+- There are some problems with viewing certain Bibles in formatted chapter view
+- This needs speed improvements. Not sure why it has lag responding to commands.
+
 Latest change in version 19.4:
 Merged Oliver's new PR.  Thanks to Oliver.
 - Add number buttons to remote control
