+++ patches.txt
@@ -1,34 +1,40 @@
+(19.5, "folder", "db")
 (18.7, "folder", "gui")
 (18.6, "folder", "htmlResources/fonts")
 (18.3, "folder", "htmlResources/css")
 (18.3, "folder", "htmlResources/js")
-(19.4, "file", "Languages.py")
-(19.4, "file", "TextCommandParser.py")
+(19.5, "file", "htmlResources/js/common.js")
+(19.5, "file", "db/Highlight.py")
+(19.5, "file", "BibleBooks.py")
+(19.5, "file", "BiblesSqlite.py")
+(19.5, "file", "Languages.py")
+(19.5, "file", "README.md")
+(19.5, "file", "TextCommandParser.py")
+(19.5, "file", "UniqueBibleAppVersion.txt")
+(19.5, "file", "gui/MainWindow.py")
+(19.5, "file", "gui/MoreConfigOptions.py")
+(19.5, "file", "gui/NoteEditor.py")
+(19.5, "file", "htmlResources/css/dark.css")
+(19.5, "file", "htmlResources/css/default.css")
+(19.5, "file", "htmlResources/js/dark.js")
+(19.5, "file", "htmlResources/js/default.js")
+(19.5, "file", "latest_changes.txt")
+(19.5, "file", "main.py")
+(19.5, "file", "patches.txt")
 (19.4, "file", "gui/RemoteControl.py")
-(19.4, "file", "patches.txt")
-(19.4, "file", "latest_changes.txt")
-(19.4, "file", "UniqueBibleAppVersion.txt")
-(19.3, "file", "gui/MainWindow.py")
-(19.2, "file", "BiblesSqlite.py")
 (19.2, "file", "ThirdParty.py")
 (19.2, "file", "gui/CentralWidget.py")
 (19.2, "file", "install.sh")
-(19.2, "file", "main.py")
-(19.1, "file", "BibleBooks.py")
 (19.0, "file", ".gitignore")
 (19.0, "file", "ToolsSqlite.py")
-(18.9, "file", "gui/MoreConfigOptions.py")
 (18.9, "file", "gui/Downloader.py")
 (18.9, "file", "themes.py")
 (18.9, "file", "patches.txt")
 (18.9, "file", "latest_changes.txt")
 (18.9, "file", "UniqueBibleAppVersion.txt")
-(18.9, "file", "htmlResources/css/dark.css")
-(18.9, "file", "htmlResources/css/default.css")
 (18.9, "file", "gui/WebEngineView.py")
 (18.7, "file", "gui/ImportSettings.py")
 (18.7, "file", "gui/MorphDialog.py")
-(18.7, "file", "gui/NoteEditor.py")
 (18.7, "file", "gui/TabWidget.py")
 (18.7, "file", "gui/WebEngineViewPopover.py")
 (18.7, "file", "gui/YouTubePopover.py")
@@ -41,8 +47,6 @@
 (18.6, "file", "htmlResources/fonts/sileotsr.ttf")
 (18.5, "file", "gui.py")
 (18.3, "file", "UniqueBible.app.pdf")
-(18.3, "file", "htmlResources/js/dark.js")
-(18.3, "file", "htmlResources/js/default.js")
 (18.0, "file", "NoteSqlite.py")
 (17.6, "file", "htmlResources/theText.js")
 (15.9, "file", "translations.py")
